Updated the wolfssl_example so its just benchmark and test. Updated the readme.txt.
